+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `Annotation` tags section within the Annotation Details.
 - Added undoable action for editing annotations on the sample details.
 - Allowed users to remove `Annotation` tags from the Annotation Details.
+- Added `AnnotationClassBalancingStrategy` class, useable in selection.
 
 ### Changed
 - Updated button text to "View sample" in annotation details panel for better clarity.